Moved InsultiBotAlexaResource to a package structure matching CompliBotAlexaResource
